+++ packages/shared-components-mezzarino/tsup.config.ts
@@ -1,13 +1,15 @@
 import { defineConfig, Options } from "tsup";
 
 export default defineConfig((options: Options) => ({
-  treeshake: true,
-  splitting: true,
   entry: ["src/**/*.tsx"],
-  format: ["esm"],
+  format: ["cjs"],
   dts: true,
   minify: true,
-  clean: true,
   external: ["react"],
+  esbuildOptions(options) {
+    options.banner = {
+      js: "'use client'",
+    }
+  },
   ...options,
 }));